[TOKN-61] Improve code organization and scalability
